requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.354
langchain-openai==0.0.5
python-multipart==0.0.6
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.112.0
pydantic==2.5.1
requests==2.31.0
caldav==1.3.6
icalendar==5.0.11